requirements.txt
pytest
pytest-selenium
selenium
Python-dotenv